Rectify build file location changes and moved test files to new location
