+++ .eslintrc.json
@@ -2,7 +2,6 @@
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaVersion": 2017,
-        "project": "tsconfig.json",
         "sourceType": "module"
     },
     "plugins": ["@typescript-eslint/eslint-plugin"],
