+++ CHANGELOG.md
@@ -7,27 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-### [v0.61.0]
+### Changed
 
-### [v0.60.0]
+-   Build file location from `dist/server.js` to `dist/src/server.js`
 
-### [v0.59.0]
+### Fixed
 
-### [v0.58.0]
-
-### [v0.57.0]
-
-### [v0.56.0]
-
-### [v0.55.0]
-
-### [v0.54.0]
-
-### [v0.53.0]
-
-### [v0.52.0]
-
-### [v0.51.0]
+-   Bug in heroku build not running due to build file location changes
 
 ### [v0.50.0]
 
@@ -224,17 +210,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Internal, dev-only test routes and methods.
 -   API documentation temporarily set up as front-end page to be served, for ease of testing by frontend.
 
-[unreleased]: https://github.com/Project-Kampong/kampong-backend/compare/v0.61.0...HEAD
-[v0.61.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.61.0
-[v0.60.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.60.0
-[v0.59.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.59.0
-[v0.58.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.58.0
-[v0.57.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.57.0
-[v0.56.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.56.0
-[v0.55.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.55.0
-[v0.54.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.54.0
-[v0.53.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.53.0
-[v0.52.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.52.0
+[unreleased]: https://github.com/Project-Kampong/kampong-backend/compare/v0.51.0...HEAD
 [v0.51.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.51.0
 [v0.50.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.50.0
 [v0.48.0]: https://github.com/Project-Kampong/kampong-backend/releases/tag/v0.48.0
