+++ package.json
@@ -2,7 +2,7 @@
     "name": "kampong-backend",
     "version": "1.0.0",
     "description": "Kampong Backend API",
-    "main": "dist/server.js",
+    "main": "dist/src/server.js",
     "engines": {
         "node": "14.x",
         "yarn": "1.x"
