+++ tsconfig.json
@@ -14,5 +14,5 @@
         "incremental": true,
         "moduleResolution": "node"
     },
-    "include": ["src", "jest.config.ts"]
+    "exclude": ["src/tests", "jest.config.ts"]
 }
